Fixing the dateformat issue + some other tweaks
Date formats can be edited in profiles. The field defaults to the current
dateformat set (either the default or the user's), but will only be saved in
the database if it differs from $defaultdateformat.

Edit Profile (editprofile.php) has been cleaned up and now does less absurd
things. For one, it doesn't pass along your password in plaintext any more!

The formatting of the file has also been cleaned up a bit and should
hopefully be less absurd.